shrubbery: make `,` optional in `@` arguments
In normal mode, groups within `()` need to have an explicit comma
separator, because that's a style that we want to enforce (and provide
good error messages about) for function arguments. When trasitioning
from text mode to normal shrubbery mode, however, it's nice to be able
to write `@form(` and then use block styling for the arguments. Using
`@(form:` is workable, but ugly, and having text after the open
parenthesis triggers a different indentation rule for the closing
parenthesis. Meanwhile, for something like `doc`, which needs
arguments after a region that is naturally block-like, I find myself
frequently missing commas.

Lifting the requirement that arguments are separated by `,` for
`@`-prefixed calls is a simple adjustment that is easy to explain and
makes a lot of things look better and easier to type.