add `syntax_class.together` and refine `fields` clause form
Change `fields` so that it declares the repetion depth of a field, and
check that patterns are consistent with the declaration. The
`syntax_class.together` form is for declaring syntax classes that
refer to themselves and each other, in which case `fields` is
required.

Also, allow pattern clauses to be "inlined" as `syntax_class`
alternatives, even when other clauses are provided. The `pattern`
clause is still available, because it could be useful in the expansion
of a syntax class clause macro.