change `Name` to recognize dotted names; reorganize syntax meta
The main goal of this change is to help macros match and manipulate
dotted sequences. A pattern escape `$(x :: Name)` matches an
identifier, operator, or sequence of components separated by `.` ---
where all components are identifiers, except that the last one can be
a parenthesized operator. The `IdentifierName` syntax class is
similar, but it recognizes only identifiers, which is along the lines
of forms like `def` and `namespace` that bind only identifiers and not
operators.

These sequence classes have the kind `~sequence`, which means they can
be used in the middle of a group. They match a spliced sequence of
terms indepedent of binding. They match eagerly, just as a consequence
of how matching is eager in general.

After matching a dotted path, the next step might be too look up
something about its binding. Better support for that step leads to a
larger set of changes in this commit. The former
`Syntax.equal_binding` function is now `syntax_meta.equal_binding` and
provided by `rhombus/meta` instead of `rhombus`; comparing bindings
may need to consult compile-time values to resolve dotted names, which
doesn't work in run-time contexts. Stepping back, a funciton like
`free-identifier=?` is used in two different ways in Racket: by a
macro to decide how to expand, or by a tool that processes expanded
code to inspect bindings. The latter doesn't really apply in Rhombus,
so we focus on the former and put it in `rhombus/meta`.

A related cleanup is to rename `Syntax.error` to `syntax_meta.error`.
The different capitalization of `Syntax` versus `syntax_meta` may not
be obvious at first glance, but it makes things more consistent
overall (e.g., even at compile time, `Syntax.error` could not be used
as `.error` after a syntax-object expression).

To use a function like `syntax_meta.equal_binding` or
`syntax_meta.value` (formerly `Syntax.meta_value`) for a space other
than the expression space, some compile-time value needs to refer to
the space. So, there are now things like `expr_meta.space`,
`bind_meta.space`, and the `reflection` meta-clause for the `space`
form.

More goodies: `Term`, `Group`, `Name`, and `IdentifierName` are now
all annotations as well as syntax classes. For example,
`Syntax.meta_value` expects a value satisfying `Name` as its argument.