add automatic functional-update `with` dot-syntax to a class
Example:

 class Posn(x, y)
 Posn(1, 2).with(x = 10)
 // prints `Posn(10, 2)`

The general syntax is `obj.with(field_id = expr, ...)` for an `obj`
with static information for class C. The result is obtained by calling
the constructor for C with arguments from `obj`'s field values, except
that for each mentioned `field_id`, the result of the corresponding
`expr` is used, instead.

A `class` declaration can replace `with` or fill it in for a custom
constructor by using the `dot` class-clause form.