add `with` for functional update of objects
Example:

 class Posn(x, y)
 Posn(1, 2) wit (x = 10)
 // prints `Posn(10, 2)`

The general syntax is `obj_expr with (field_id = expr, ...)`. For an
`obj_expr` with static information for class C, the result is obtained
by calling the constructor for C with arguments from `obj`'s field
values, except that for each mentioned `field_id`, the result of the
corresponding `expr` is used, instead.

A `class` declaration can replace `with` or fill it in for a custom
constructor by using the new `udpdate` class-clause form, which is
similar to `expression` or `annotation`.

The `with` form requires static information from `obj_expr` to
identify a class. Otherwise, a syntax error is reported.

When a class is abstract or has a custom constructor, then there's no
default implementation for `with`, but `update` can supply one.