add `expr.merge`
The `expr.merge` form allows an identifier to be an expression form as
well as definition form, etc. A special form is needed because those
all use the default space.

Anothert possibile direction would be to use different binding spaces,
but that comes with its own trade-offs. The common case, by far, is
that an identifier works in only one of those, and using the default
space for these kinds of bindings improves error reporting.