add `Array.of_length` annotation and reducer
Mainly to replace the `~length` form of `Array`.  My rationale is that
reducer macros should consume a finite number of terms in the presence
of the `each` shorthand; the existing parsing hack makes me very
uncomfortable.  Having a grouping in hand also makes it possible to
add a `~fill` option, as in `for/vector`.