draw and gui: use `enum` for symbol enumerations
Some smalle backward-incompatible changes: `~style` for many GUI
elements becomes `~styles`, because it's a list of symbols such as
ones in the `Window.Style` enumeration. The previous approach was to
call the list a "style" and the enumeration `Window.StyleSymbol`, but
it's better to have "style" consistently correspond to one symbol.