draw and gui: use `enum` for symbol enumerations
Also, as a small but backward-incompatible changes, change the
`~style` keyword argument for many GUI elements to `~styles`. Those
arguments are a list of symbols, such as ones in the `Window.Style`
enumeration. (The previous approach was to call the list a "style" and
the enumeration `Window.StyleSymbol`, but it's better to have "style"
consistently correspond to an individual symbol.)