`error` and `error.message`
Change `error` to take its `who` argument via the keyword `~who`, and
add support for message details and clauses. Clause constructors like
`error.val` and `error.annot` build the label--value pairs that go at
the end of an error message. The `error.message` function can be used
to just construct the message text that `error` puts into an
exception.

This commit does not change the non-message content of exceptions, but
the clause protocol is meant to accommodate and adapt to future
changes.