generalize list binding patterns to support multiple splices
Allow multiple `...` repetitions and `&` splices in a list pattern. In
general, matching such patterns can require a search, and there can be
multiple matches, but it's the same in syntax patterns, and the search
implemented here is similarly greedy (i.e., longer matches to earlier
subpatterns are preferred).

The low-level binding protocol was not quite right to support a
backtracking search, so much of this commit is adding one more element
to the low-level protocol. Adapting all the predefined binding forms
was tedious. While making the protocol yet more complex is worrying,
supporting searches seems worthwhile. Also, the set of predefined
forms is rich enough --- especially with `&&`, `||` and `where`
patterns --- to expect that most binding macros are (and will be)
derived from the predefined ones, instead of written directly with the
low-level protocol.

One difference between syntax matching and list matching is that
there's no notion of splciig bindings forms, while syntax matching has
splicing syntax classes. Almost the same effect can be achieved with
list patterns using `&`. Example:

```
```