add `definitely` epression and binding form
The `definitely` form is a destructor to go with the `maybe`
constructor. As an expression, `definitely(expr)` produces the same
result as `expr`, except that it errors if that result is `#false`.
And if `expr` has static information from `maybe(annot)`, then
`definitely(expr)` has the static information of `annot`. The
`definitely(bind)` binding form matches only non-`#false` values that
match `bind`. Using `definitely` is the same as writing
`expr :: annot` or `bind :: annot`, except that you don't have to
write out `annot`.

The commit also adjusts `match` to propagate static information from
the target expression to binding patterns, which it should have done
anyway, but is especially needed to make `definitely` work as
intended.